Generate more events
Fixed the issue with not generating the expected number of events. Added in a constraint, such that the W is constrained to be above threshold for the reaction.

Changed the way that the scattered electron is determined: done by generating flat Q2 and flat E_scattered distributions. Then determines the theta and 4 momentum.